refactor: split remapper.py
